Attempt to fix circular dependency, if that's really the issue
